--- frontend/app/host/worker/webWorker.ts
@@ -1,7 +1,0 @@
-// 
-
-export const createWorker = (worker: any) => {
-    const code = worker.toString();
-    const blob = new Blob(["(" + code + ")()"]);
-    return new Worker(URL.createObjectURL(blob));
-};
